+++ dart/geobase/lib/src/utils/bounds_builder.dart
@@ -9,10 +9,10 @@ import 'dart:math' as math;
 import 'package:meta/meta.dart';
 
 import '/src/codes/coords.dart';
+import '/src/coordinates/base/bounded.dart';
 import '/src/coordinates/base/box.dart';
 import '/src/coordinates/base/position.dart';
 import '/src/coordinates/base/position_series.dart';
-import '/src/vector_data/model/bounded/bounded.dart';
 
 /// A helper class to calculate bounds for a set of points and other bounds.
 ///