+++ dart/geobase/lib/src/vector_data/model/geometry/geometry.dart
@@ -9,6 +9,7 @@ import 'dart:typed_data';
 import 'package:meta/meta.dart';
 
 import '/src/codes/geom.dart';
+import '/src/coordinates/base/bounded.dart';
 import '/src/coordinates/projection/projection.dart';
 import '/src/coordinates/reference/coord_ref_sys.dart';
 import '/src/vector/content/geometry_content.dart';
@@ -17,7 +18,6 @@ import '/src/vector/encoding/binary_format.dart';
 import '/src/vector/encoding/text_format.dart';
 import '/src/vector/formats/geojson/geojson_format.dart';
 import '/src/vector/formats/wkb/wkb_format.dart';
-import '/src/vector_data/model/bounded/bounded.dart';
 
 /// A base interface for geometry classes.
 ///