+++ dart/geobase/lib/vector_data.dart
@@ -19,7 +19,6 @@
 /// Usage: import `package:geobase/vector_data.dart`
 library vector_data;
 
-export 'src/vector_data/model/bounded/bounded.dart';
 export 'src/vector_data/model/feature/feature.dart';
 export 'src/vector_data/model/feature/feature_builder.dart';
 export 'src/vector_data/model/feature/feature_collection.dart';