+++ dart/geobase/CHANGELOG.md
@@ -13,6 +13,7 @@ NOTE: Version 1.0.0 currently under development (1.0.0-dev.0).
 
 🛠 Refactoring:
 * [Normalize longitude with inclusive limit on east 180 degrees #210](https://github.com/navibyte/geospatial/issues/210)
+* [Move Bounded to coordinates and implement it by PositionSeries #212](https://github.com/navibyte/geospatial/issues/212)
 
 ## 0.6.0